feat(cli): remove HashiCorp checkpoint telemetry transport
Phase B of the HashiCorp-telemetry removal (002, closes the core ask of
issue #48): the CLI no longer makes any request to
checkpoint-api.hashicorp.com.

- delete checkpoint.ts (ReportRequest, ReportParams, post(), BASE_URL)
- relocate the preserved identity utilities getUserId/getProjectId
  (+ private getId/homeDir) to identity.ts — Sentry scope tagging
  depends on them
- errors.ts: drop the fire-and-forget report() side effect from the
  Errors.Internal/External/Usage factories; Sentry integration stays
- sendTelemetry becomes a signature-preserving no-op in telemetry.ts;
  the 7 call sites keep compiling and will be reconnected to Sentry v10
  metrics (with consent gating) in the next change
- delete the checkpoint nock test
