+++ packages/@cdktn/commons/src/errors.ts
@@ -1,34 +1,18 @@
 // Copyright (c) HashiCorp, Inc
 // SPDX-License-Identifier: MPL-2.0
-import { ReportParams, ReportRequest } from "./checkpoint";
-import { DISPLAY_VERSION } from "./version";
 import * as Sentry from "@sentry/node";
-
-// Errors that will emit telemetry events
-async function report(command: string, payload: Record<string, any>) {
-  const reportParams: ReportParams = {
-    command,
-    product: "cdktn",
-    version: `${DISPLAY_VERSION}`,
-    dateTime: new Date(),
-    payload,
-  };
-
-  await ReportRequest(reportParams);
-}
 
 type ErrorType = "Internal" | "External" | "Usage";
 export function IsErrorType(error: any, type: ErrorType): boolean {
   return error && error.__type === type;
 }
 
-function reportPrefixedError(type: ErrorType, command: string) {
+function reportPrefixedError(type: ErrorType, _command: string) {
   return (
     message: string,
     originalError: Error = new Error(),
     context?: Record<string, any>,
   ) => {
-    report(command, { ...context, message, type });
     const err: any = new Error(`${type} Error: ${message}`);
     Object.entries(context || {}).forEach(([key, value]) => {
       err[key] = value;
