+++ packages/@cdktn/commons/src/index.ts
@@ -1,16 +1,17 @@
 // Copyright (c) HashiCorp, Inc
 // SPDX-License-Identifier: MPL-2.0
 
-export * from "./checkpoint";
 export * from "./construct-maker-target";
 export * from "./config";
 export * from "./debug";
 export * from "./environment";
 export * from "./errors";
 export * from "./gradle";
+export * from "./identity";
 export * from "./logging";
 export * from "./module-schema";
 export * from "./provider-schema";
+export * from "./telemetry";
 export * from "./terraform-module";
 export * from "./terraform";
 export * from "./util";
